+++ examples/slack/run.sh
@@ -3,7 +3,7 @@
 SERVER_ADDR="$1"
 
 rm -rf target && mkdir target
-cp -R concord.yml _main.json target/
+cp -R concord.yml target/
 
 cd target && zip -r payload.zip ./* > /dev/null && cd ..