+++ runner/src/main/java/com/walmartlabs/concord/runner/engine/SlackTask.java
@@ -1,7 +1,9 @@
 package com.walmartlabs.concord.runner.engine;
 
-import com.walmartlabs.concord.common.Task;
+import com.walmartlabs.concord.sdk.Constants;
+import com.walmartlabs.concord.sdk.Context;
 import com.walmartlabs.concord.sdk.InjectVariable;
+import com.walmartlabs.concord.sdk.Task;
 import io.takari.bpm.api.BpmnError;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -19,6 +21,14 @@ public class SlackTask implements Task {
     @Inject
     public SlackTask(RpcClient rpc) {
         this.rpc = rpc;
+    }
+
+    @Override
+    public void execute(Context ctx) throws Exception {
+        String instanceId = (String) ctx.getVariable(Constants.Context.TX_ID_KEY);
+        String channelId = (String) ctx.getVariable("channelId");
+        String text = (String) ctx.getVariable("text");
+        call(instanceId, channelId, text);
     }
 
     public void call(@InjectVariable("txId") String instanceId, String channelId, String text) throws Exception {